+++ docs/CHANGELOG.rst
@@ -14,6 +14,7 @@ Added
 -----
 - Add ``move_to_collection`` to ``Data`` viewset
 - Report registration failure in ``ProcessTestCase``
+- Add a pseudo Python process to serve as a template
 
 
 ===================